Add idempotency token auto fill
Adds support for service models which specify a shape member is the
idempotency token to be auto filled with a random UUID if the member's
value is not already set.

This replaces the needs for customizations per service per request when
an idempotency token should always be provided, but optionally set by the
user.