aws/request: Add support for logging and stopping handler lists
Adds support for setting a logger on `request.Handlerlist` used by each
service client. This allows you to provide a debug logging function
which is called for each iteration in the handler list. A default
logging func is also provided, `LogHandlerListItems` which will log the
handler used, if named, and the current state of the request's Error.

Adds support for stopping a `request.HandlerList`'s iteration after a
request handler has been executed. A default `HandlerListStopOnError`
func is provided to stop iterating a handler list's request handlers if
any request error occurs. This is added to the default Build handler
list. To stop if the request fails to build.