private/signer/v4: Resign requests which are too old
Requests that are delayed to be sent, or are retried so many times that
the original date the request was signed with is past the age a
signature can be, the request needs to be resigned.  This prevents
issues for request being retried many times, and finally succeding but
encounter an expired signature.

Fix #486