private/signer/v4: Include Content-Length in signature
Removes the content-length header from the HTTP request if its value is
zero. This would of already been performed by the Go http.Client, but
the SDK needs to do this earlier to ensure the content-lenght header
does not get included in the signature.

This change also ensures the bode's content-length is signed if set to a
value other than 0.

Fix #625