Update testing's go vet usage to not fail on tip
Updates to Makefile so the go vet doesn't fail on go tip. Go tip changed
the flag usaged to ignore example vet errors that was added in Go 1.6.

Also adds a Dockerfile so the SDK can be tested against Go tip locally
without having to wait for travis to run remotely.