aws/request: Fix data race caused by resusing request header map value
If the header map values are reused between retry attempts it is
possible for a race codition to occur where transporter is still
referencing the header values when the request is being retried and
resigned.

This change updates the request copy to do a deep copy for headers and
URL. In addition the request fields not used by the SDK are dropped.

Fix #679