service/ec2: Add support for new params for Spot Fleet
Amazon Elastic Compute Cloud

The new 'type' parameter in the RequestSpotFleet API indicates whether a
Spot fleet will only 'request' the target capacity or also attempt to
'maintain' it. When you want to 'maintain' a certain target capacity,
Spot fleet will place the required bids to meet this target capacity,
and enable you to automatically replenish any interrupted instances.
When you simply 'request' a certain target capacity, Spot fleet will
only place the required bids but will not attempt to replenish Spot
instances if capacity is diminished, nor will it submit bids in
alternative Spot pools if capacity is not available. By default, this
parameter is set to 'maintain'. - The DescribeSpotFleetRequests API now
returns two new parameters: the 'fulfilledCapacity' of a Spot fleet to
indicate the capacity that has been successfully launched, and the
'type' parameter to indicate whether the fleet is meant to 'request' or
'maintain' capacity.